# openapi.yaml
openapi: 3.0.0
info:
  title: Main Collection
  version: 1.0.0
  description: Contains the Postman docs for all of our api/endpoionts.
servers:
  - url: http://localhost:3000
paths:
...
  /api/poll/{pollId}/vote:
    post:
      summary: Vote in Poll
      responses: {}
      tags:
        - Polls
      security:
        - BearerAuth: []
      description: >-
        ## **POST /api/poll/\[id\]/vote** – Cast a Vote


        Create a vote for a poll. Each user can only vote **once** per poll.


        ### **Authentication**


        - Header: `Authorization: Bearer`
            

        ### **URL Parameters**


        | Parameter | Type | Required | Description |

        | --- | --- | --- | --- |

        | `id` | number | Yes | Poll ID to vote on. |


        ### **Request Body**


        | Field | Type | Required | Description |

        | --- | --- | --- | --- |

        | `optionId` | number | Yes | The ID of the poll option to vote for. |


        **Example Request Body**


        ```

        {  "optionId": 5}

         ```

        ### **Success Response (201 Created)**


        ```

        {  

        "message": "Vote cast successfully",  

        "vote": 
             {    "id": 123,    "pollId": 45,    "optionId": 5,    "userId": 12  }
        }

         ```

        ### **Error Responses**


        | Status | Condition | Response |

        | --- | --- | --- |

        | 400 | Missing pollId or optionId | `{ "error": "Missing required
        fields" }` |

        | 400 | Invalid pollId or optionId format | `{ "error": "Invalid ID
        format" }` |

        | 400 | User already voted | `{ "error": "User has already voted in this
        poll" }` |

        | 401 | Missing/invalid token | `{ "error": "Unauthorized" }` / `{
        "error": "Invalid token" }` |

        | 403 | User banned | `{ "error": "User is banned" }` |

        | 404 | Poll not found | `{ "error": "Poll not found" }` |

        | 500 | Server error | `{ "error": "Failed to cast vote" }` |
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                optionId:
                  type: integer
            example:
              optionId: 1
    parameters:
      - name: pollId
        in: path
        required: true
        schema:
          type: string
        example: '{{pollId}}'
    get:
      summary: Get Vote
      responses: {}
      tags:
        - Polls
      security:
        - BearerAuth: []
      description: >-
        ## **GET /api/poll/\[id\]/vote** – Check User Vote


        Retrieve the current authenticated user's vote for a poll. Returns
        `null` if the user has not voted.


        ### **Authentication**


        - Header: `Authorization: Bearer`
            

        ### **URL Parameters**


        | Parameter | Type | Required | Description |

        | --- | --- | --- | --- |

        | `id` | number | Yes | Poll ID to check vote for. |


        ### **Success Response (200 OK)**


        - If user has voted:
            

        ```

        {  
             "userVote": 
                {    "id": 123,    "pollId": 45,    "optionId": 5,    "userId": 12  }
        }

         ```

        - If user has **not voted**:
            

        ```

        {  "userVote": null}

         ```

        ### **Error Responses**


        | Status | Condition | Response |

        | --- | --- | --- |

        | 400 | Invalid pollId | `{ "error": "Invalid poll ID" }` |

        | 401 | Missing/invalid token | `{ "error": "Unauthorized" }` / `{
        "error": "Invalid token" }` |

        | 403 | User banned | `{ "error": "User is banned" }` |

        | 404 | Poll not found | `{ "error": "Poll not found" }` |

        | 500 | Server error | `{ "error": "Failed to check user vote" }` |
      requestBody:
        content:
          application/json:
            schema:
              type: object
    patch:
      summary: Update Vote
      responses: {}
      tags:
        - Polls
      security:
        - BearerAuth: []
      description: >-
        ## **PATCH /api/poll/\[id\]/vote** – Update Vote


        Change an existing vote to a different poll option.


        ### **Authentication**


        - Header: `Authorization: Bearer`
            

        ### **URL Parameters**


        | Parameter | Type | Required | Description |

        | --- | --- | --- | --- |

        | `id` | number | Yes | Poll ID to update vote for. |


        ### **Request Body**


        | Field | Type | Required | Description |

        | --- | --- | --- | --- |

        | `optionId` | number | Yes | The new option ID to vote for. |


        **Example Request Body**


        ```

        {  "optionId": 6}

         ```

        ### **Success Response (200 OK)**


        ```

        {  
             "message": "Vote updated successfully",  
             "vote": 
             {    "id": 123,    "pollId": 45,    "optionId": 6,    "userId": 12  }
        }

         ```

        ### **Error Responses**


        | Status | Condition | Response |

        | --- | --- | --- |

        | 400 | Missing pollId or optionId | `{ "error": "Missing required
        fields" }` |

        | 400 | Invalid pollId or optionId | `{ "error": "Invalid ID format" }`
        |

        | 400 | User has not voted yet | `{ "error": "User has not voted in this
        poll" }` |

        | 401 | Missing/invalid token | `{ "error": "Unauthorized" }` / `{
        "error": "Invalid token" }` |

        | 403 | User banned | `{ "error": "User is banned" }` |

        | 404 | Poll not found | `{ "error": "Poll not found" }` |

        | 500 | Server error | `{ "error": "Failed to update vote" }` |
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                optionId:
                  type: integer
            example:
              optionId: 2
    delete:
      summary: Remove Vote
      responses: {}
      tags:
        - Polls
      security:
        - BearerAuth: []
      description: >-
        ## **DELETE /api/poll/\[id\]/vote** – Remove Vote


        Remove the current user's vote from a poll (like “unvoting”).


        ### **Authentication**


        - Header: `Authorization: Bearer`
            

        ### **URL Parameters**


        | Parameter | Type | Required | Description |

        | --- | --- | --- | --- |

        | `id` | number | Yes | Poll ID to remove vote from. |


        ### **Success Response (200 OK)**


        ```

        {  "message": "Vote removed successfully"}

         ```

        ### **Error Responses**


        | Status | Condition | Response |

        | --- | --- | --- |

        | 400 | Invalid pollId | `{ "error": "Invalid poll ID" }` |

        | 401 | Missing/invalid token | `{ "error": "Unauthorized" }` / `{
        "error": "Invalid token" }` |

        | 403 | User banned | `{ "error": "User is banned" }` |

        | 404 | Poll or vote not found | `{ "error": "Poll not found" }` / `{
        "error": "Vote not found" }` |

        | 500 | Server error | `{ "error": "Failed to remove vote" }` |
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
security:
  - BearerAuth: []